Create Future Work list
Made a list on Dr. Wilson's request that gives an outline of some possible features that we wanted to add to our application but never had the time or ability to. This is meant to give some poor souls in a few years an idea of how this project can be improved